requirements.txt
Django==5.2.*
psycopg2-binary==2.9.*
Pillow==10.0.*
boto3==1.34.*
django-storages==1.14.*
